rul-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pandas>=2.0,<2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
tensorflow-cpu==2.15.1